+++ sea-query-binder/src/sqlx_sqlite.rs
@@ -8,104 +8,105 @@ impl<'q> sqlx::IntoArguments<'q, sqlx::sqlite::Sqlite> for SqlxValues {
             use sqlx::Arguments;
             match arg {
                 Value::Bool(b) => {
-                    args.add(b);
+                    let _ = args.add(b);
                 }
                 Value::TinyInt(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::SmallInt(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::Int(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::BigInt(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::TinyUnsigned(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::SmallUnsigned(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::Unsigned(i) => {
-                    args.add(i);
+                    let _ = args.add(i);
                 }
                 Value::BigUnsigned(i) => {
-                    args.add(i.map(|i| <i64 as std::convert::TryFrom<u64>>::try_from(i).unwrap()));
+                    let _ = args
+                        .add(i.map(|i| <i64 as std::convert::TryFrom<u64>>::try_from(i).unwrap()));
                 }
                 Value::Float(f) => {
-                    args.add(f);
+                    let _ = args.add(f);
                 }
                 Value::Double(d) => {
-                    args.add(d);
+                    let _ = args.add(d);
                 }
                 Value::String(s) => {
-                    args.add(s.map(|s| *s));
+                    let _ = args.add(s.map(|s| *s));
                 }
                 Value::Char(c) => {
-                    args.add(c.map(|c| c.to_string()));
+                    let _ = args.add(c.map(|c| c.to_string()));
                 }
                 Value::Bytes(b) => {
-                    args.add(b.map(|b| *b));
+                    let _ = args.add(b.map(|b| *b));
                 }
                 #[cfg(feature = "with-chrono")]
                 Value::ChronoDate(d) => {
-                    args.add(d.map(|d| *d));
+                    let _ = args.add(d.map(|d| *d));
                 }
                 #[cfg(feature = "with-chrono")]
                 Value::ChronoTime(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-chrono")]
                 Value::ChronoDateTime(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-chrono")]
                 Value::ChronoDateTimeUtc(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-chrono")]
                 Value::ChronoDateTimeLocal(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-chrono")]
                 Value::ChronoDateTimeWithTimeZone(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-time")]
                 Value::TimeDate(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-time")]
                 Value::TimeTime(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-time")]
                 Value::TimeDateTime(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-time")]
                 Value::TimeDateTimeWithTimeZone(t) => {
-                    args.add(t.map(|t| *t));
+                    let _ = args.add(t.map(|t| *t));
                 }
                 #[cfg(feature = "with-uuid")]
                 Value::Uuid(uuid) => {
-                    args.add(uuid.map(|uuid| *uuid));
+                    let _ = args.add(uuid.map(|uuid| *uuid));
                 }
                 #[cfg(feature = "with-rust_decimal")]
                 Value::Decimal(decimal) => {
                     use rust_decimal::prelude::ToPrimitive;
-                    args.add(decimal.map(|d| d.to_string()));
+                    let _ = args.add(decimal.map(|d| d.to_string()));
                 }
                 #[cfg(feature = "with-bigdecimal")]
                 Value::BigDecimal(big_decimal) => {
                     use bigdecimal::ToPrimitive;
-                    args.add(big_decimal.map(|d| d.to_string()));
+                    let _ = args.add(big_decimal.map(|d| d.to_string()));
                 }
                 #[cfg(feature = "with-json")]
                 Value::Json(j) => {
-                    args.add(j.map(|j| *j));
+                    let _ = args.add(j.map(|j| *j));
                 }
                 #[cfg(feature = "with-ipnetwork")]
                 Value::IpNetwork(_) => {
@@ -118,6 +119,10 @@ impl<'q> sqlx::IntoArguments<'q, sqlx::sqlite::Sqlite> for SqlxValues {
                 #[cfg(feature = "postgres-array")]
                 Value::Array(_, _) => {
                     panic!("Sqlite doesn't support array arguments");
+                }
+                #[cfg(feature = "postgres-vector")]
+                Value::Vector(_) => {
+                    panic!("Sqlite doesn't support vector arguments");
                 }
             }
         }